fix: force tab trigger slot to be non-lazy
Seems like in React 19 there were some significant changes to the way
React Server Components worked and this is the root cause of the issue
we're seeing with tabs. After inspecting the tabs in the React dev
extension, I noticed that the tabs that were missing from the tab list
were there they were just lazy and weren't being rendered by react while
the other tabs had been rendered and had been realised into proper
elements.

This lead me to a few issues which seemed to relate to this problem, but
most notably this one: https://github.com/radix-ui/primitives/issues/3776
on the radix ui repo which is very close to our issue. Applying the
suggested fix to force React turn the children into real elements seems
to work for us too, so this is a sufficient workaround for now.

There's most likely a future step we need to take to fix this properly,
but this works for now and it seems like we're not alone in having
issues related to lazy components and the way they are now handled in
React 19 with RSC, so maybe some of our depedencies will provide
fixes (e.g. React or radix ui) or guidance.